ready_cache: use `futures::task::AtomicWaker` for cancelation (#669)
This replaces the use of `tokio::sync::Notify` with
`futures::task::AtomicWaker`. `Notify` requires the `Notified` future to
be held in order to remain interested in the notification. Dropping the
future returned by `Notify::notified` cancels interest in the
notification. So, the current code using `Notify` is incorrect, as the
`Notified` future is created on each poll and then immediately dropped,
releasing interest in the wakeup.

We could solve this by storing the `Notify::notified` future in the
`Pending` future, but this would be a bit of a pain, as the `Notified`
future borrows the `Notify`. I thought that it was a nicer alternative
to just rewrite this code to use `AtomicWaker` instead.

Also, `AtomicWaker` is a much simpler, lower-level primitive. It simply
stores a single waker that can wake up a single task. `Notify` is
capable of waking multiple tasks, either in order or all at once,
which makes it more complex.