util: add `Wrap`
This is middleware that wraps a service with both pre and post
functions. It offers two flavors, one which is purely for passing shared
state from the pre function to the post, and the other of which allows
for asynchronous pre/post functions, mapping the request type, and
short-circuiting so the inner service isn't called.