chore: note `Budget` breaking change in changelog
`Budget` is now a trait in the 0.5 release. this is a breaking change relative to the 0.4 release, where it was a concrete [struct](https://docs.rs/tower/0.4.13/tower/retry/budget/struct.Budget.html).

this commit updates the changelog to characterize this as a breaking change, rather than an additive change.